chore: upgrade PHPStan from level 6 to level 7
Resolved all 484 L7 errors across 68 files with zero baseline:
- Added DecimalMath wrapper (app/Support/DecimalMath.php) for type-safe
  bc* operations, replacing 295 raw bcadd/bcsub/bcmul/bcdiv/bccomp calls
- Added @var annotations for Model::find() union type narrowing (149 errors)
- Cast auth()->id() to int, e() to string, sum() to string (41 misc fixes)

All 339 tests pass. No behavioral changes.

+++ app/Filament/Resources/Fulfillment/ShippingOrderExceptionResource/Pages/ViewShippingOrderException.php
@@ -154,7 +154,7 @@ class ViewShippingOrderException extends ViewRecord
                 ->action(function (ShippingOrderException $record): void {
                     $record->status = ShippingOrderExceptionStatus::Resolved;
                     $record->resolved_at = now();
-                    $record->resolved_by = Auth::id();
+                    $record->resolved_by = (int) Auth::id();
                     $record->save();
 
                     Notification::make()
